code_company_backend/requirements.txt
flask
requests
schedule
openai
python-dotenv
supabase